Changed laser hockey to hockey env and added option of multiple rounds
